requirements.txt
pymupdf
pillow
pytesseract
pdf2image
pyyaml
openpyxl
python-docx
